+++ luck-report-server/luck-report-plugins/luck-report-infra/src/main/java/com/luck/report/infra/modules/cache/service/ReportCache.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  ******************************************************************************/
-package com.luck.report.core.cache;
+package com.luck.report.infra.modules.cache.service;
 
 
 import java.util.Set;