+++ luck-report-server/luck-report-plugins/luck-report-infra/src/main/java/com/luck/report/infra/modules/cache/service/ReportCacheKeyResolver.java
@@ -1,4 +1,4 @@
-package com.luck.report.core.cache;
+package com.luck.report.infra.modules.cache.service;
 
 /**
  * 缓存键生成策略