+++ luck-report-server/luck-report-plugins/luck-report-infra/src/main/java/com/luck/report/infra/modules/servlet/provider/RequestInfoProvider.java
@@ -1,4 +1,4 @@
-package com.luck.report.web.provider;
+package com.luck.report.infra.modules.servlet.provider;
 
 import java.util.Enumeration;