+++ luck-report-server/luck-report-web/src/main/java/com/luck/report/web/service/ReportExportService.java
@@ -16,7 +16,7 @@ import com.luck.report.core.export.pdf.PdfProducer;
 import com.luck.report.core.export.word.high.WordProducer;
 import com.luck.report.core.model.Report;
 import com.luck.report.web.constant.ReportConstants;
-import com.luck.report.web.provider.RequestInfoProvider;
+import com.luck.report.infra.modules.servlet.provider.RequestInfoProvider;
 import com.luck.report.web.utils.UrlParameterUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;