+++ luck-report-server/luck-report-web/src/main/java/com/luck/report/web/utils/UrlParameterUtils.java
@@ -16,7 +16,7 @@
 package com.luck.report.web.utils;
 
 import com.luck.report.web.exception.ReportDesignException;
-import com.luck.report.web.provider.RequestInfoProvider;
+import com.luck.report.infra.modules.servlet.provider.RequestInfoProvider;
 
 import java.io.UnsupportedEncodingException;
 import java.net.URLDecoder;